refactor: extract telegram pending edit handler
